Update entireReceiver___arcade_drive_TL_servo.ino
We changed a few of the pins that are set on the motor controller. The left joystick is accurately controlling the motor but the right one is not responding at all so the receiver code needs to be checked again. The switch values collection needs to be checked again as it is not accurately commanding the weapon motor